% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{charge_corr}
\alias{charge_corr}
\title{Correct misassigned and unassigned charges for monoisotopic peaks}
\usage{
charge_corr(data, mz_tol, scan_tol, min_iso)
}
\arguments{
\item{data}{a \code{data.table} output of \code{\link[mzcharge]{charge_spectrafile}} function.}

\item{mz_tol}{max allowed deviation in m/z between centroids in neighboring scans (ppm).}

\item{scan_tol}{max allowed gap between scans to consider them nighbors.}

\item{min_iso}{min number of isotopologues used for initial charge assignment.}
}
\value{
a \code{data.table} containing monoisotopic peaks with corrected charge.
}
\description{
Correct misassigned and unassigned charges for monoisotopic peaks
}
