% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{charge_singlespec}
\alias{charge_singlespec}
\title{Convert profile spectra to centroided spectra with assigned charge and
position in isotopic envelope.}
\usage{
charge_singlespec(
  spectrum,
  min_width,
  min_intensity,
  mz_tol,
  intensity_tol,
  max_charge,
  num_iso
)
}
\arguments{
\item{spectrum}{a 2-column numeric matrix with mz [,1] and intenity [,2].}

\item{min_width}{a size of sliding window for centroiding.}

\item{min_intensity}{minimal intensity for centroiding.}

\item{mz_tol}{max mz deviation between isotopologues (ppm).}

\item{intensity_tol}{max deviation between measured intensity and predicted
using averagine model.}

\item{max_charge}{max charge to assign.}

\item{num_iso}{max number of isotopologues to search for.}
}
\value{
A \code{data.frame}
}
\description{
\code{charge_singlespec} converts single profile spectrum into
centroided spectrum using local maximum across \code{min_width} points.
Assigment of charge and isotopic position is based on mz distance between
isotopologues and deviation of measured intensity from predicted by
averagine model.
For high performance, \code{charge_singlespec} is implemented in C++ using
\href{http://www.rcpp.org}{Rcpp}.
}
