% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{xic_monomz}
\alias{xic_monomz}
\title{Extract ion current for monoisotopic mz/charge pair}
\usage{
xic_monomz(data, query_mz, query_charge, mz_tol)
}
\arguments{
\item{data}{a \code{data.table} output of \code{\link[mzcharge]{charge_corr}} function.}

\item{query_mz}{query mz.}

\item{query_charge}{query charge.}

\item{mz_tol}{max allowed deviation from \code{query_mz} in ppm.}
}
\value{
a \code{data.table}.
}
\description{
Extract ion current for monoisotopic mz/charge pair
}
